Update template to use u64 instead of u32
Some solutions require the bigger storage, so use it by default.

+++ template.rs
@@ -18,6 +18,6 @@ fn main() -> Result<(), io::Error> {
     Ok(())
 }
 
-fn solution(input: &str) -> u32 {
+fn solution(input: &str) -> u64 {
     0
 }
